First pass, everything compiles, tests pass, under Spark 1.3.0

+++ tile-generation/src/test/scala/com/oculusinfo/tilegen/datasets/SchemaTypeUtilitiesTestSuite.scala
@@ -30,13 +30,14 @@ import java.lang.{Long => JavaLong}
 import java.lang.{Double => JavaDouble}
 
 import org.apache.spark.SharedSparkContext
-import org.apache.spark.sql.catalyst.types.ArrayType
+import org.apache.spark.sql.types.ArrayType
 import org.scalatest.FunSuite
 
-import org.apache.spark.sql.{Row, StructField, StructType}
-import org.apache.spark.sql.{BooleanType, StringType, TimestampType}
-import org.apache.spark.sql.{ByteType, ShortType, IntegerType, LongType}
-import org.apache.spark.sql.{FloatType, DoubleType}
+import org.apache.spark.sql.Row
+import org.apache.spark.sql.types.{StructField, StructType}
+import org.apache.spark.sql.types.{BooleanType, StringType, TimestampType}
+import org.apache.spark.sql.types.{ByteType, ShortType, IntegerType, LongType}
+import org.apache.spark.sql.types.{FloatType, DoubleType}
 import org.apache.spark.sql.catalyst.expressions.GenericRow
 
 import scala.collection.mutable.ArrayBuffer
@@ -298,13 +299,13 @@ class SchemaTypeUtilitiesSparkTestSuite extends FunSuite with SharedSparkContext
 
 		assert(3 === data.schema.fields.size)
 		assert("a" == data.schema.fields(0).name)
-		assert(IntegerType == data.schema.fields(0).dataType)
+		assert(LongType == data.schema.fields(0).dataType)
 		assert("b" == data.schema.fields(1).name)
-		assert(IntegerType == data.schema.fields(1).dataType)
+		assert(LongType == data.schema.fields(1).dataType)
 		assert("c" == data.schema.fields(2).name)
-		assert(IntegerType == data.schema.fields(2).dataType)
+		assert(LongType == data.schema.fields(2).dataType)
 
-		val converter = calculatePrimitiveConverter(IntegerType, classOf[Double])
+		val converter = calculatePrimitiveConverter(LongType, classOf[Double])
 		val augmentFcn: Array[Any] => Any = row =>
 			{
 				val c = converter(row(0))
@@ -314,11 +315,11 @@ class SchemaTypeUtilitiesSparkTestSuite extends FunSuite with SharedSparkContext
 		val augmentedData = addColumn(data, "d", DoubleType, augmentFcn, "c", "a")
 		assert(4 === augmentedData.schema.fields.size)
 		assert("a" == augmentedData.schema.fields(0).name)
-		assert(IntegerType == augmentedData.schema.fields(0).dataType)
+		assert(LongType == augmentedData.schema.fields(0).dataType)
 		assert("b" == augmentedData.schema.fields(1).name)
-		assert(IntegerType == augmentedData.schema.fields(1).dataType)
+		assert(LongType == augmentedData.schema.fields(1).dataType)
 		assert("c" == augmentedData.schema.fields(2).name)
-		assert(IntegerType == augmentedData.schema.fields(2).dataType)
+		assert(LongType == augmentedData.schema.fields(2).dataType)
 		assert("d" == augmentedData.schema.fields(3).name)
 		assert(DoubleType == augmentedData.schema.fields(3).dataType)
 
@@ -328,9 +329,9 @@ class SchemaTypeUtilitiesSparkTestSuite extends FunSuite with SharedSparkContext
 		for (i <- 1 to 10) {
 			val row = collectedData(i-1)
 			assert(4 === row.size)
-			val a = row(0).asInstanceOf[Int]
-			val b = row(1).asInstanceOf[Int]
-			val c = row(2).asInstanceOf[Int]
+			val a = row(0).asInstanceOf[Long]
+			val b = row(1).asInstanceOf[Long]
+			val c = row(2).asInstanceOf[Long]
 			val d = row(3).asInstanceOf[Double]
 
 			assert(i === a)
